requirements.txt
firebase-admin
google-api-python-client
google-cloud-vision
mcp3008
picamera
RPi.GPIO
spidev
